// src/services/signupInvitations/searchByTokenUserSignupInvitation/types.ts
export interface ISearchByTokenUserSignupInvitationResponse {
  userIdentification: string;
  userName: string;
  email: string;
  id?: string;
  invitationId?: string;
  businessUnit?: string;
  phoneNumber?: string;
  userAccountId?: string;
  dateStart?: string;
  dateEnd?: string;
  requestingUser?: string;
  status?: string;
  password?: string;
}